Move CompressedChunk to it's own file
